Move process denial to generate appeal
